requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
click==7.1.2
cycler==0.10.0
Django==3.0.6
django-cors-headers==3.3.0
djangorestframework==3.11.0
joblib==0.15.1
kiwisolver==1.2.0
matplotlib==3.2.1
mysqlclient==1.4.6
nltk==3.5
numpy==1.18.4
pandas==1.0.3
pkg-resources==0.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.5.14
scikit-learn==0.23.0
scipy==1.4.1
six==1.14.0
sklearn==0.0
soupsieve==2.0.1
sqlparse==0.3.1
threadpoolctl==2.0.0
tqdm==4.46.0
